Reorganizes components to be in appropriate folders and fixes their imports for files they are used in
